remove bundle wrapper from message send
Instead of defaulting to not wrapping, this goes a very explicit route and removes the bundle wrapper entirely.

The setRemoteEndpoint call was also removed from the sender example as it is not necessary and potentially misleading. In the current API, Senders send to a fixed host and port, setting the remote host/port on a message is used on the receiving side to keep track of where message came from. In turn, setting it from the sending side doesn't functionally do anything and give the wrong impression that it controls where the message is being sent to.