avoid deprecated API for OS version when possible
The Gestalt functions have been deprecated since 10.8 (and are said to incorectly report 10.10 as 10.9 though I was unable to reproduce on 10.10.3) and API was added to NSProcessInfo to replace it. Though the API wasn't made public until 10.10 / iOS 8, it is available on earlier releases (possibly backported in more recent SDKs due to Gestalt being deprecated without a replacement) so the Objective-C runtime is used to to divine its availability and falls back to Gestalt. The NSProcessInfo API is also available on iOS and is a bit cleaner than the string parsing, so if the API is available, iOS will use it as well.