better handle ivar memory leak
Second bite of the cherry - I didn't think about it, but mPlayerVideoOutput is an ivar and it needs to be handled a bit differently than a leaking local variable. Technically, the existing change made in 557a6d2 will do the trick as the player item retains the video output, but it is ugly and not idiomatic.