request security exception for insecure content
Like Flickr, it appears as though api.instragram.com also needs an exception to disable the perfect forward secrecy requirement.

It is worth notihng, the sample is using some older gl matrix stuff and does not build as is, but I tested this by commenting that portion out and just focusing on the networking bits.