release CFURL after use
createCfUrl will return an owned reference, so we need to dispose of it when we are done with it. This is a bit tricky since it was being used in the error handling conditional so for simplicity I just changed the exception to use the original file path string, similar but not exactly the same thing.